pyproject.toml
[project]
name = "medrag"
version = "0.1.0"
description = "Medical RAG (Retrieval-Augmented Generation) application for querying medical PDF documents"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jupyter>=1.1.1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "torch>=2.8.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "chromadb>=0.4.22",
    "pypdf>=4.0.1",
    "sentence-transformers>=2.2.2",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.2",
    "faiss-cpu>=1.7.4",
    "openai>=1.12.0",
    "tqdm>=4.66.1",
    "pydantic>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88
